% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{caramella}
\alias{caramella}
\title{Caramella mesh}
\description{
A triangular mesh of a manifold representing a candy used as toy geometry.This is a MESH.2.5D object created with \code{create.MESH.2.5D}.
}
