% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdaPDE.objects.R
\name{create.FEM.basis}
\alias{create.FEM.basis}
\title{Create a FEM basis}
\usage{
create.FEM.basis(mesh)
}
\arguments{
\item{mesh}{A \code{MESH2D} or \code{MESH.2.5D} object representing the domain triangulation. See \link{create.MESH.2D}, \link{create.MESH.2.5D}.}
}
\value{
A  \code{FEMbasis} object. This contains the \code{mesh}, along with some additional quantities:

if \code{class(mesh) == MESH2D}
	\item{\code{order}}{Either "1" or "2". Order of the Finite Element basis.} 
	\item{\code{nbasis}}{Scalar. The number of basis.} 
	\item{\code{transf_coord}}{An object containing 4 vectors of length #triangles. The for of them encode the tranformation matrix [diff1x diff2x; diff1y diff2y] that transforms the nodes of the reference triangle to the nodes of the i-th triangle.}
	\item{\code{detJ}}{A vector of length #triangles. The ith element contains the determinant of the transformation from the reference triangle to the nodes of the i-th triangle. It's values is also the double of the area of each triangle of the basis.}
if \code{class(mesh) == MESH.2.5D}
	\item{\code{order}}{Either "1" or "2". Order of the Finite Element basis.}
	\item{\code{nbasis}}{Scalar. The number of basis.}
}
\description{
Sets up a Finite Element basis. It requires a triangular mesh, a \code{MESH2D} or a \code{MESH.2.5D} object, as input. 
The basis' functions are globally continuos surfaces, that are polynomials once restricted to a triangle in the mesh. 
Linear if (\code{order = 1}) in the input \code{mesh} and quadratic if (\code{order = 2}) in the input \code{mesh}
Finite Element are currently implemented.
}
\examples{
## Creates a simple triangulated domain with a concavity; this is a MESH2D object  
mesh<-create.MESH.2D(nodes=rbind(c(0, 0), c(0, 1), c(0.5, 0.5), c(1, 1), c(1, 0)),
segments=rbind(c(1, 2), c(2, 3), c(3, 4), c(4, 5), c(5, 1)), order=1)
## Plot it
plot(mesh)                   
## Creates the basis
FEMbasis = create.FEM.basis(mesh)
}
\seealso{
\code{\link{create.MESH.2D}}, \code{\link{create.MESH.2.5D}}
}
