% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh.2D.R
\name{create.MESH.2.5D}
\alias{create.MESH.2.5D}
\title{Create a \code{MESH.2.5D} object from the connectivty matrix and nodes locations}
\usage{
create.MESH.2.5D(nodes, triangles, order = 1)
}
\arguments{
\item{nodes}{A nnodes x 3 matrix specifying the locations of each node}

\item{triangles}{A ntriangles x 3*order matrix specifying the indices of the nodes in each triangle}

\item{order}{Either "1" or "2". Order of the Finite Element basis default is order = 1}
}
\value{
A \code{MESH.2.5D} object
}
\description{
Create a \code{MESH.2.5D} object from the connectivty matrix and nodes locations
}
\examples{
#read the matrix nodes and triangles from file
nodes = read.table(file="mynodes.csv",header=F,sep=",")
triangles = read.table(file="mytriangles.csv",header=F,sep=",")
mesh = create.MESH.2.5D(nodes,triangles)
}
