% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdaPDE.locator.R
\name{evaluate.grid.FEM}
\alias{evaluate.grid.FEM}
\title{Evaluate a FEM object at a grid of point locations}
\usage{
eval.FEM(FEM, length_interval_x=25, length_interval_y=NULL,length_interval_z=NULL, CPP_CODE=TRUE)
}
\arguments{
\item{FEM}{A \code{FEM} object to be evaluated.}

\item{CPP_CODE}{Boolean. If \code{TRUE} the computation relies on the C++ implementation of a Visibility Walk Algorithm (Devillers et al. 2001). This usually ensures a fast computation. In case of the 2D manifold in a 3D space or a 3D volume, only the C++ method is available.}

\item{length_interval_x.}{An integer specifying the length of the interval on the x axis of the grid of points where the FEM object should be evaluated. Default value is 25.}

\item{length_interval_y.}{An integer specifying the length of the interval on the y axis of the grid of points where the FEM object should be evaluated. If no value is specified, the length is assumed to be equal to the \code{length_interval_x} parameter.}

\item{length_interval_z.}{An integer specifying the length of the interval on the z axis of the grid of points where the FEM object should be evaluated. If no value is specified, the length is assumed to be equal to the \code{length_interval_x} parameter.}
}
\value{
An object of class \code{FEMevaluated} with the following output:
\item{\code{evaluated_values}}{A matrix of numeric evaluations of the \code{FEM} object. The element of the matrix corresponds to the locations of the elements of the meshgrid generated.}
\item{\code{meshgrid}}{The meshgrid generated using the specified intervals on the 3 axis.}
\item{\code{x}}{A vector of length \code{length_interval_x} containing the x coordinate for each point of the meshgrid generated.}
\item{\code{y}}{A vector of length \code{length_interval_y} containing the y coordinate for each point of the meshgrid generated.}
\item{\code{z}}{A vector of length \code{length_interval_z} containing the z coordinate for each point of the meshgrid generated.}
}
\description{
It evaluates a FEM object at a specified grid of point locations. It is needed for doing 3D plots like isosurface plot, slice plot and slicecontour plot.
}
\references{
Devillers, O. et al. 2001. Walking in a Triangulation, Proceedings of the Seventeenth Annual Symposium on Computational Geometry
}
\seealso{
\code{\link{isosurfaces.FEMevaluated}},\code{\link{slices.FEMevaluated}} and \code{\link{slicecontours.FEMevaluated}}
}
