% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{ICA}
\alias{ICA}
\title{Internal Carotid Artery mesh}
\description{
A triangular mesh of a manifold representing the geometry of an internal carotid artery. This is a MESH.2.5D object created with \code{create.MESH.2.5D}.
}
