% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdaPDE.plots.R
\name{isosurfaces.FEMevaluated}
\alias{isosurfaces.FEMevaluated}
\title{Isosurface Plot of a FEMevaluated object}
\usage{
isosurfaces.FEMevaluated(FEMevaluated,levels,...)
}
\arguments{
\item{FEMevaluated}{A \code{FEMevaluated} object.}

\item{levels}{The levels at which the isosurfaces are generated. If levels is a vector, isosurfaces are generated for each level. The value of the levels has to be in the range of the possible \code{evaluated_values} of the \code{FEMevaluated} object.}

\item{...}{Arguments representing  graphical options to be passed to \link[rgl]{plot3d}.}
}
\description{
Isosurface plot of a \code{FEMevaluated} object, generated by the function \code{evaluate.grid.FEM}. This method is useful for plotting volumetric mesh (\code{MESH.3D}) objects.
}
\examples{
## Upload a tetrahedral mesh and plot it
data(sphere3D)
plot(sphere3D)
##Define the number of nodes
nnodes=sphere3D$nnodes
## Create a linear Finite Element basis
FEMbasis = create.FEM.basis(sphere3D)
## Define a function as expansion of this basis
coeff=numeric(nnodes)
for(i in 0:nnodes-1){
	coeff[i+1] <- 0.5*sqrt(15/pi)*sphere3D$nodes[3*i+1]*sphere3D$nodes[3*i+2]
}
FEM_object<- FEM(coeff, FEMbasis)
##Create the grid of evaluated points
FEMevaluated=evaluate.grid.FEM(FEM_object)
##Plot the isosurfaces
isosurfaces.FEMevaluated(FEMevaluated,levels=seq(-0.45,0.45,by=0.15))
}
\seealso{
\code{\link{evaluate.grid.FEM}}
}
