% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdaPDE.plots.R
\name{plot.FEM}
\alias{plot.FEM}
\title{Plot a \code{FEM} object}
\usage{
\method{plot}{FEM}(x, num_refinements, ...)
}
\arguments{
\item{x}{A \code{FEM} object.}

\item{num_refinements}{A natural number specifying how many bisections should be applied to each triangular element for
plotting purposes. This functionality is useful where a discretization with 2nd order Finite Element is applied. This parameter can be specified only when a planar mesh (\code{MESH2D}) object is plotted.}

\item{...}{Arguments representing graphical options to be passed to \link[rgl]{plot3d}.}
}
\description{
Three-dimensional plot of a \code{FEM} object, generated by \code{FEM} or returned by \code{smooth.FEM.basis}, \code{smooth.FEM.PDE.basis},
\code{smooth.FEM.PDE.sv.basis} or \code{smooth.FEM.FPCA}. 
If the \code{mesh} of the \code{FEMbasis} component is of class \code{MESH2D} both the 3rd axis and the color represent the value of the coefficients for the Finite Element basis expansion (\code{coeff} component of the \code{FEM} object). If the \code{mesh} is of class \code{MESH.2.5D}, the color of each triangle represent the mean value of the coefficients for the Finite Element basis expansion(\code{coeff}).
}
\examples{
## Upload a triangular mesh and plot it
data("mesh.2D.rectangular")
plot(mesh.2D.rectangular)
## Create a linear Finite Element basis
FEMbasis = create.FEM.basis(mesh.2D.rectangular)
## Define a sinusoidal function as expansion of this basis and plot it
coeff <- sin(mesh.2D.rectangular$nodes[,1])*cos(mesh.2D.rectangular$nodes[,2])
FEM_object<- FEM(coeff, FEMbasis)
plot(FEM_object)
}
\seealso{
\code{\link{image.FEM}}, \code{\link{FEM}}
}
