% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh.2D.R
\name{plot.MESH.2.5D}
\alias{plot.MESH.2.5D}
\title{Plot a triangular mesh}
\usage{
\method{plot}{MESH.2.5D}(mesh, node_values = NULL)
}
\arguments{
\item{mesh}{A \code{MESH.2.5D} object}

\item{node_values}{(optional) A vector with nodal values of the function to be rapresented on the mesh, if it is \code{NULL} only the triangulation will be plotted, otherwhise a colormap will be produced}
}
\description{
Plot a triangular mesh
}
\examples{
library(fdaPDE)
data(caramella)
plot.MESH.2.5D(caramella)
library(fdaPDE)
data(caramella)
### GENERATE SOME RANDOM DATA, SEE FOR EXAMPLE test_manifold1.R IN THE TEST DIRECTORY ###
output = smooth.FEM.basis( observations = data, FEMbasis = FEMbasis, lambda = 1,GCV = FALSE, CPP_CODE = TRUE)
plot.MESH.2.5D(FEMbasis$mesh, output_CPP$fit.FEM$coeff)
}
