% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdaPDE.plots.R
\name{plot.MESH.3D}
\alias{plot.MESH.3D}
\title{Plot a MESH.3D object}
\usage{
\method{plot}{MESH.3D}(mesh, ...)
}
\arguments{
\item{mesh}{A \code{MESH.3D} object generated by \code{create.MESH.3D} or \code{second.order.MESH.3D}.}

\item{...}{Arguments representing graphical options to be passed to \link[graphics]{par}.}
}
\description{
Plot the triangulation of a mesh \code{MESH.3D} object, generated by \code{create.MESH.3D}. The points indicates the mesh nodes.
}
\examples{
##Load a MESH.3D object
data(sphere3D)
##Plot the triangulation of the object
plot.MESH.3D(sphere3D)
}
