% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdaPDE.smoothing_R.R
\name{R_elementProperties}
\alias{R_elementProperties}
\title{Compute some properties for each triangular element of the mesh}
\usage{
R_elementProperties(mesh)
}
\arguments{
\item{mesh}{A \code{MESH2D} mesh object representing the triangular mesh. This can be created with  \code{\link{create.MESH.2D}}.}
}
\value{
A list with the following variables:
\item{\code{transf_coord}}{An object containing 4 vectors of length #triangles. The for of them encode the tranformation matrix [diff1x diff2x; diff1y diff2y] that transforms the nodes of the reference triangle to the nodes of the i-th triangle.}
\item{\code{detJ}}{A vector of length #triangles. The ith element contains the determinant of the transformation from the reference triangle to the nodes of the i-th triangle. It's values is also the double of the area of each triangle of the basis.}
}
\description{
It computes some quantities associated to the linear map that transforms the ith triangle in the reference triangular element. 
These are used for the computation of the integrals necessary to build the mass and stiffness matrix.
This method is implemented and available only for \code{MESH2D} mesh objects.
}
