% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdaPDE.smoothing_R.R
\name{R_eval.FEM}
\alias{R_eval.FEM}
\title{Evaluate a FEM object at a set of locations}
\usage{
R_eval.FEM(FEM, locations)
}
\arguments{
\item{FEM}{A \code{FEM} object to be evaluated}

\item{locations}{A 2-columns matrix with the spatial locations where the FEM object should be evaluated}
}
\value{
A matrix of numeric evaluations of the \code{FEM} object. Each row indicates the location where the evaluation has been taken, the column indicates the 
function evaluated.
}
\description{
Only executed when the function \code{smooth.FEM.basis} is run with the option \code{CPP_CODE} = \code{FALSE}. It evaluates a FEM object at the specified set of locations.
}
\seealso{
\code{\link{R_eval.FEM.basis}}
}

