% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh.2D.R
\name{second.order.MESH.2.5D}
\alias{second.order.MESH.2.5D}
\title{Double the order of a fist order Finite Element mesh by adding middle points to each side of the triangles in the triangulation}
\usage{
second.order.MESH.2.5D(mesh,bc=NULL)
}
\arguments{
\item{mesh}{an object of class 'MESH.2.5D' that is the starting mesh of order 1}

\item{bc}{A vector specifying the indices of the nodes on which boundary conditions are applied.}
}
\value{
If no boundaries conditions are passed:
\item{\code{mesh}}{An object of class 'MESH.2.5D' with the mesh of order 2.} Otherwhise a \code{list} with parameters:
\item{\code{mesh}}{An object of class 'MESH.2.5D' with the mesh of order 2.}
\item{\code{bc_index}}{An update of the vector specifying the indices of the nodes on which boundary conditions are applied.}
}
\description{
Double the order of a fist order Finite Element mesh by adding middle points to each side of the triangles in the triangulation
}
\examples{
#Loading mesh hub, a \\code{MESH.2.5D} object of order=1 contained in the package
data(hub)
#Apply the function to generate a surface mesh of order 2
hub_order2 = second.order.MESH.2.5D(hub)
}
\seealso{
\code{\link{create.MESH.2.5D}}
}
