% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdaPDE.plots.R
\name{slicecontours.FEMevaluated}
\alias{slicecontours.FEMevaluated}
\title{Slice Contour Plot of a FEMevaluated object}
\usage{
slicecontours.FEMevaluated(FEMevaluated,xs=0,ys=NULL,zs=NULL,...)
}
\arguments{
\item{FEMevaluated}{A \code{FEMevaluated} object.}

\item{xs}{Vector specifying the position in x where the slices(planes) has to be drawn. The values of \code{evaluated_values} of the \code{FEMevaluated} object are projected on these slices as contours.}

\item{ys}{Vector specifying the position in y where the slices(planes) has to be drawn. The values of \code{evaluated_values} of the \code{FEMevaluated} object are projected on these slices as contours.}

\item{zs}{Vector specifying the position in z where the slices(planes) has to be drawn. The values of \code{evaluated_values} of the \code{FEMevaluated} object are projected on these slices as contours.}

\item{...}{Arguments representing  graphical options to be passed to \link[rgl]{plot3d}.}
}
\description{
Slice contour plot of a \code{FEMevaluated} object, generated by the function \code{evaluate.grid.FEM}. This method is useful for plotting volumetric mesh (\code{MESH.3D}) objects.
}
\examples{
## Upload a tetrahedral mesh and plot it
data(sphere3D)
plot(sphere3D)
##Define the number of nodes
nnodes=sphere3D$nnodes
## Create a linear Finite Element basis
FEMbasis = create.FEM.basis(sphere3D)
## Define a function as expansion of this basis
coeff=numeric(nnodes)
for(i in 0:nnodes-1){
	coeff[i+1] <- 0.5*sqrt(15/pi)*sphere3D$nodes[3*i+1]*sphere3D$nodes[3*i+2]
}
FEM_object<- FEM(coeff, FEMbasis)
##Create the grid of evaluated points
FEMevaluated=evaluate.grid.FEM(FEM_object)
##Plot the contour on the slices
slicecontours.FEMevaluated(FEMevaluated,xs=c(-0.5,0,0.5),ys=0,NAcol="blue")
}
\seealso{
\code{\link{evaluate.grid.FEM}}
}
