% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{sphereData}
\alias{sphereData}
\title{SphereData}
\description{
A dataset with information about the connectivity matrix and the nodes locations of a sphere geometry. It containes:
\itemize{
	\item nodes. A #nodes-by-3 matrix specifying the locations of each node.
	\item triangles. A #triangles-by-3*order matrix specifying the indices of the nodes in each triangle.
   }
This dataset is used to create a \code{MESH.2.5D} object with the function \code{create.MESH.2.5D}
}
