% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdaPDE.plots.R
\name{plot.MESH.2.5D}
\alias{plot.MESH.2.5D}
\title{Plot a MESH.2.5D object}
\usage{
\method{plot}{MESH.2.5D}(mesh)
}
\arguments{
\item{mesh}{A \code{MESH.2.5D} object generated by \code{create.MESH.2.5D} or \code{second.order.MESH.2.5D}.}
}
\description{
Plot the triangulation of a mesh MESH2D object, generated by \code{create.MESH.2.5D} or \code{second.order.MESH.2.5D}.
}
\examples{
#Load a \\code{MESH.2.5D} object
data(caramella
#Plot the triangulation of the object
plot.MESH.2.5D(caramella)
}
