% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdaPDE.plots.R
\name{plot.MESH.2.5D}
\alias{plot.MESH.2.5D}
\title{Plot a MESH.2.5D object}
\usage{
\method{plot}{MESH.2.5D}(mesh, ...)
}
\arguments{
\item{mesh}{A \code{MESH.2.5D} object generated by \code{create.MESH.2.5D} or \code{second.order.MESH.2.5D}.}

\item{...}{Arguments representing graphical options to be passed to \link[graphics]{par}.}
}
\description{
Plot the triangulation of a mesh \code{MESH.2.5D} object, generated by \code{create.MESH.2.5D} or \code{second.order.MESH.2.5D}. The points indicates the mesh nodes.
}
\examples{
##Load a MESH.2.5D object
data(candy)
##Plot the triangulation of the object
plot.MESH.2.5D(candy)
}
